Install pytest modules for more flexibility
- `pytest-asyncio` has been added for testing async code
- `pytest-xdist` added for parallelising tests
- `pytest-github-actions-annotate-failures` for automatically generating github annotations

The introduction of these modules changes how we run pytest. Coverage and results
are now output primarily as XML alongside the terminal versions.
Tracking coverage is not enabled by default to allow for fast testing with parallelisation

Tests are run with 'new' first (i.e. not in the result cache), and a test taking
more than a second is reported as a potential issue