Switch Sonar reporting to an active mode
Rather than have Sonar automatically scan the repo based on a webhook,
we instead explicitly upload the state of a request into it.
This method allows us to also upload information like the coverage report,
and the reports from the linters that Sonar is configured to use.

The ToolChain code in the tools module has been updated to output all
toolchain output to files in `reports/`.