RSS Feed Input (#39)
RSS Input System

Allows using an RSS feeds as an Input.
Events are generated whenever a new items is added to the feed (identified by GUID), and contain the contents of the RSS document.

The Input is implemented using a polling loop with a configurable interval time. Multiple feeds can be handled by Input.