Configure setuptools to work with multiple packages
This change adds support for building and installing the mewbot subpackages mewbot-core, mewbot-api and so forth.
It works by dynamically generating the setuptools configuration based on the arguments given to setup.py.

Each of the named packages will include the module `mewbot.[name]` (or just `mewbot` for the `mewbot` package), and will
pull the dependencies from the `requirements-[name].txt` file. Packages that are not `mewbot-core` will also depend
on `mewbot-core`.