tools: Configure to work with any src-dir projects
This change is the first step for introducing the concepts of both
`mebot.tools` as an external package which can be pulled in to lint
another codebase (such as a mewbot plugin), and for restructuring
the repository to have in-tree plugins.

The change adds the capability for `mewbot.tools.path` to dynamically
generate paths by looking for folders matching a certain name in a
given location. Currently it looks only in `.`, but this can easily
be expanded.

Along with these changes, the ability to filter what paths are examined
by the paths has been added, with both the `lint` and `test` tool chains
taking a list of paths to run against, and the `test` tools also
supporting a list of coverage references being selected.

This allows for easier checking of the state of a specific module.
For example

```sh
./tools/test tests/io/test_io_rss.py --cover mewbot.io.rss
```

Will run only one test file, and produce a coverage report for a single
class. This should reduce testing times when working locally on a
specific component.

Additionally, the entire module's documentation has been updated.