actions: Only lint merge files that have changed
When running in a merge request context, we only need to run the linters
against the changed files. This aligns better with the Sonar concept of
not penalising merge requests for code they have not changed, and will
allow for easier enabling of new linting tools (such as documentation
requirements).