Refactor Actions to not require knowledge of queues
This change makes each call to `Behaviour.process` and `Action.act`
less stateful by moving all references to the queue out to the bot
or runner levels of interaction.
Both of these methods now emit (in the form of an async generator)
all output events that would be created.

As many actions will not emit any events (for example for updating
state information of recording data to datastores), the ability to
`yield None` has been codified in. This allows dvelopers to ensure
they are providing an async generator.