Add 'Reply' generic events and action to the Common IO.
A InputEvent abstract class/mix-in/protocol has been added which
allows an InputEvent to create an OutputEvent which is a reply to it.

This event contains a string message, to which no formatting information
is assigned.

There are two scopes for the reply -- either the same scope as the
incoming message, or in a potentially 'narrower' scope which still
includes the sender, but fewer other people (such as a direct message).

These two modes may be equivalent.

A ReplyAction has also been added, using the `string.Template` to
format the current state into the message, making it able to combine
arbitrary IO systems with other arbitrary actions.